requirements.txt
# Core app dependencies
gradio==4.36.1
python-docx==1.1.0
docx2txt==0.8
pypdf==4.2.0
chromadb==0.4.24
sentence-transformers==2.2.2

# Data handling
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.4

# LLM clients
openai==1.35.14
google-generativeai==0.5.2

# Networking (used by Gradio internally)
httpx==0.27.0